Replace Panel_HeaderUI with HeaderUI - Moves all the HeaderUI code to Panel_HeaderUI - Fix an 'undefined' being put in the className if not defined
