+++ src/panels/PanelHeader.tsx
@@ -1,34 +1,49 @@
 import { observer } from 'mobx-react-lite'
+import React from 'react'
+import { ReactNode } from 'react'
 
-import { RevealUI } from 'src/rsuite/reveal/RevealUI'
+function correctChildrenHeight(elements: [ReactNode] | ReactNode) {
+    if (Array.isArray(elements)) {
+        return React.Children.map(elements, (child, index) => {
+            if (React.isValidElement(child)) {
+                return React.cloneElement(child, {
+                    className: `${child.props.className ? child.props.className + ' ' : ''}CSHY-panel-header-item`,
+                } as React.HTMLAttributes<HTMLElement>)
+            } else {
+                return child
+            }
+        })
+    } else if (React.isValidElement(elements)) {
+        return React.cloneElement(elements, {
+            className: `${elements.props.className ? elements.props.className + ' ' : ''}CSHY-panel-header-item`,
+        } as React.HTMLAttributes<HTMLElement>)
+    }
+}
 
-export const PanelHeaderUI = observer(function PanelHeaderUI_(p: {
-    /** @default "Options" */
-    label?: string
-    children: React.ReactNode
-}) {
+/** Re-usable Dock-Panel Header, gives a `bg-base-300` bar with a horizontal flex to put widgets in.
+ *
+ * `NOTE`: It will automatically set the height of any child widgets.
+ *
+ * Example:
+ *
+ * ```
+ * <PanelHeaderUI>
+ *      <div tw='btn btn-sm'>Hello World!<div>
+ * </PanelHeaderUI>
+ * ```
+ */
+export const PanelHeaderUI = observer(function PanelHeaderUI_(p: { children?: ReactNode }) {
+    let children = p.children
+    if (p.children) {
+        children = correctChildrenHeight(p.children)
+    }
     return (
-        <div className='cushy-panel-header'>
-            <RevealUI tooltipWrapperClassName='p-2'>
-                <div tw='btn btn-sm'>
-                    <span className='material-symbols-outlined'>settings</span>
-                    {p.label ?? 'Options'}
-                </div>
-                <div tw='min-w-96'>{p.children}</div>
-            </RevealUI>
-        </div>
-    )
-})
-
-export const PanelHeaderSmallUI = observer(function PanelHeaderSmallUI_(p: { children: React.ReactNode }) {
-    return (
-        <div className='cushy-panel-header'>
-            <RevealUI tw='w-full'>
-                <div tw='btn btn-sm w-full'>
-                    <span className='material-symbols-outlined'>settings</span>
-                </div>
-                <div tw='min-w-96'>{p.children}</div>
-            </RevealUI>
+        <div // Container
+            // This may be good to pass in the future? I don't think this really needs to/should be modifiable though.
+            // className={p.className}
+            tw={['CSHY-panel-header', 'flex select-none w-full', 'items-center p-1', 'bg-base-300']}
+        >
+            {children ?? <></>}
         </div>
     )
 })
