+++ src/panels/Panel_Import.tsx
@@ -3,15 +3,15 @@ import { observer } from 'mobx-react-lite'
 import { useSt } from '../state/stateContext'
 import { ImportAsImageUI, ImportedFileUI } from 'src/importers/FilesBeeingImported'
 import { TargetBox } from 'src/importers/TargetBox'
-import { HeaderUI } from 'src/controls/widgets/header/HeaderUI'
 import { SeparatorUI } from 'src/controls/widgets/separator/SeparatorUI'
+import { PanelHeaderUI } from './PanelHeader'
 
 export const Panel_Import = observer(function Panel_Import_(p: {}) {
     const st = useSt()
 
     return (
         <div tw='w-full h-full'>
-            <HeaderUI>
+            <PanelHeaderUI>
                 <SeparatorUI />
                 {/* Putting clear items button on the right because actions that remove things should be separated in some way from other things to prevent mis-clicks. */}
                 <div
@@ -23,7 +23,7 @@ export const Panel_Import = observer(function Panel_Import_(p: {}) {
                     <span className='material-symbols-outlined'>backspace</span>
                     Clear Items
                 </div>
-            </HeaderUI>
+            </PanelHeaderUI>
             <div tw='flex flex-col gap-3 m-3'>
                 {st.droppedFiles.map((file, ix) => (
                     <div key={file.name} tw='card card-bordered p-2 bg-base-200'>
